Integration Test
* Stop executing integrstion test with PayaraMicro but with EmbeddedAll
* move IntegrationTest from wegas-runtime to wegas-app
- wegas-runtime
