+++ wegas-app/src/test/glassfish/lib/install/applications/.gitignore
@@ -1,0 +1,1 @@
+ejb-timer-service-app.war
