+++ wegas-app/src/test/glassfish/lib/install/applications/__ds_jdbc_ra/META-INF/MANIFEST.MF
@@ -1,0 +1,13 @@
+Manifest-Version: 1.0
+Ant-Version: Apache Ant 1.6.5
+Created-By: 11.0-b15 (Sun Microsystems Inc.)
+Specification-Title: JDBC Specification
+Specification-Vendor: Sun Microsystems, Inc.,
+Specification-Version: 3.0, 4.0
+Implementation-Vendor: Sun Microsystems, Inc.,
+Implementation-Version: V3
+Implementation-Title: GlassFish
+probe-provider-class-names: com.sun.gjc.monitoring.StatementCacheProbe
+ Provider,com.sun.gjc.monitoring.SQLTraceProbeProvider,com.sun.gjc.mon
+ itoring.StatementLeakProbeProvider
+
