+++ wegas-app/src/test/resources/logback.xml
