+++ wegas-app/src/test/resources/wegas.properties
@@ -1,0 +1,58 @@
+# General #
+###########
+encoding=UTF-8
+#encoding=${project.build.sourceEncoding}
+debug=${wegas.debug}
+timestamp=${wegas.build}
+clientErrorLogging=${wegas.clientErrorLogging}
+useYUICDN=${wegas.useYUICDN}
+
+# Facebook #
+############
+facebook.appSecret=15cca164c035d21b6c2bb924f051788b
+facebook.appId=214143398685715
+facebook.redirectUrl=http://wegas.dyndns.org:8080/Wegas/FacebookServlet
+
+# AAI #
+#######
+aai.secret=
+aai.enabled=false
+aai.showbutton=false
+aai.server=
+aai.loginurl=
+
+# JCR (JackRabbit) #
+####################
+jcr.admin.username=admin
+jcr.admin.password=admin
+jcr.file.maxsize=20971520
+jcr.repository.maxsize=100000000
+jcr.repository.URI=
+
+# Security #
+############
+guestallowed=true
+
+# Mail #
+########
+mail.default_domain=wegas.test
+mail.smtp.host=
+mail.smtp.username=
+mail.smtp.password=
+mail.smtp.auth=false
+mail.smtp.starttls.enable=true
+mail.smtp.port=25
+
+# Pusher #
+##########
+pusher.appId=
+pusher.key=
+pusher.secret=
+pusher.cluster=mt1
+pusher.onlineusers_hook=false
+
+# NEO4J (NLP) #
+###############
+neo4j.server.url=
+neo4j.server.user=neo4j
+neo4j.server.password=neo4j
